nits+ CPAS fix
